+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#885](https://github.com/ericaltendorf/plotman/pull/885))
 - `supervisord` now used in Docker image.
   ([#898](https://github.com/ericaltendorf/plotman/pull/898))
+- Output same entries to plotman.log from 'plotman interactive' and ' plotman plot/archive' "daemons".
+  ([#878](https://github.com/ericaltendorf/plotman/pull/878))
 
 ## [0.5.1] - 2021-07-15
 ### Fixed